Fixed player and monster image position in the battle screen window.
Added exploreDunLvl method in battleFunctions. this method randomly generates a integer and will perform different functions based on the integer. Currently the functions are find nothing, find treasure chest, find monster.

Added loadMonster method in battleFunctions. This method generates a random number based on the monInfo array index to select a random monster. The random inxex number is then used to set the variables in batMonData object. This object is used for the random monster encounters.

Added code to call to reload inventory after finding an item.

Added html code for a button to explore dungeon. In the js code a on click funtion is linked to this button. The on click code calls the exploreDunlvl method in the battleFunctions object.