Added a function in exploreDunlvl method to either drop and item or gold. If gold is dropped then it called the method updateGold to update the gold value in the inventory tab.
Added a variable in js code called crntlyexplor and set it to 0. This var will be set to 1 when the explore dungeon level button is pressed so the button cannot be repeatedly clicked to call the exploreDunLvl method mulitple times. The exploreDunLvl will set this variable to 0 once it has perform its functions.