Added html code to display player and monster status bars underneath images. Each bar is a div element which will be manipulated from the game.js code.
In game.js added a one click event for the html button element with id atkSlash. the click event checks if a battle is already triggered and if player turn is up. If true then it will set player turn to 1 and call the battleFunctions.meleeAtkSlash method.

The battleFunctions.meleeAtkSlash method calucates the players damage done to the monster and changes the batMonData curHp to minus the damage. If the curHp of the monster is less than or equal to 0 then the method battleFunctions.clearBatMonData will be called to clear the status of the monster for use on the next encounter.