Added code the method monAtk and monTurnTimer to call each other to create a loop of wait timer and monster attacked. Added if statements to both methods to quit the loop if the monster is dead.
Added code to the method meleeAtkSlask to pick a random number between 1 and 20 and then add the player hit bonus to see if the player hits or monster dodges the attack. Added a if statement inside to run the code if his and run code if misses. If miss the code will print out a messege about monster dodging.

Added a new property to the playerInfo object called hit, this is the players hit bonus used to calculate the hit chance on attacks.

Added new code in the loadPlayerStats method to insert hit bonus text into the stat tab of the menu.

Added element to the html code for the loadPlayerStats to insert hit chance variable.