Added two audio html tags for exploring dungeon music and battle music.
Added 2 audio files in the assets/audio folder.

Added a new object dungeonFunctions with a method dungeionMusic to play or pause the audio file in the audi tag with id exploringMusic.

Added new method in battleFunctions called battleMusic to play or pause the audio in the audio tag with id battleMusic.