Added two methods to the game,js code. menuFunctions.playerInvAddItem is called to add an item to the player inventory and menuFunctions.playerAddCount is called to add to the count of an existing item in the player inventory.
Added a new array method invCheckLoop to loop through player inventory to see if the item from chest already exists. Returns false if not in invenotry and returns the player inventory index if it does.

Added code in the exploreDungeon method to call on the above mentioned methods.
The new code should now update player inventory in the menu and battle menu withe items and the quantities.

In game.js itemData array properties were changed from objects to arrays.